Added a popup in onboarding

+++ client/js/user-onboard.js
@@ -1,51 +1,127 @@
 class UserOnboard {
 
-	constructor() {
-
-		this.page = window.page;
-	}
-
 	static async setup() {
 
 		if(document.querySelector('.setup-stages'))
 			document.querySelector('.setup-stages').remove();
 
-		const onboard = new UserOnboard();
+		if(!UserOnboard.instance)
+			UserOnboard.instance = new UserOnboard();
 
-		await onboard.load();
-		await Storage.set('newUser', {});
+		await UserOnboard.instance.load();
+	}
+
+	constructor() {
+
+		this.page = window.page;
+		this.stages = [];
 	}
 
 	async load() {
 
-		const container = document.createElement('div');
+		this.stages = [];
 
-		container.classList.add('setup-stages');
-
-		container.innerHTML = `<a href="${demo_url}" target="_blank">View Demo</a>`;
+		const stageLoads = [];
 
 		for(const stage of UserOnboard.stages.values()) {
 
 			const stageObj = new stage();
 
-			container.appendChild(stageObj.container);
-			await stageObj.load();
+			this.stages.push(stageObj);
+
+			stageLoads.push(stageObj.load());
 		}
+
+		await Promise.all(stageLoads);
+
+		if(this.stages.every(stage => stage.isCompleted))
+			return await Storage.delete('newUser');
+
+		if(document.querySelector('main .setup-stages'))
+			document.querySelector('main .setup-stages').remove();
+
+		document.querySelector('main').appendChild(this.container);
+
+		this.loadWelcomeDialogBox();
+	}
+
+	get container() {
+
+		if(this.containerElement)
+			return this.containerElement;
+
+		const container = this.containerElement = document.createElement('div');
+
+		container.classList.add('setup-stages');
+
+		container.innerHTML = `<a href="${demo_url}" target="_blank">View Demo</a>`;
+
+		for(const stage of this.stages)
+			container.appendChild(stage.container);
 
 		container.insertAdjacentHTML('beforeend', '<span class="close">Skip</span>');
 
 		container.querySelector('.close').on('click', async () => {
 
 			container.remove();
+
 			await Storage.delete('newUser');
 		});
 
-		if(container.querySelectorAll('.completed').length == 4) {
+		return container;
+	}
 
-			await Storage.delete('newUser');
+	async loadWelcomeDialogBox() {
+
+		if(this.stages.some(stage => stage.isCompleted))
+			return;
+
+		const newUser = await Storage.get('newUser');
+
+		if(newUser.skipWelcomeDialogBox == true)
+			return;
+
+		if(!this.dialogBox) {
+
+			this.dialogBox = new DialogBox();
+
+			this.dialogBox.container.classList.add('user-onboarding-welcome');
 		}
 
-		document.querySelector('main').appendChild(container);
+		this.dialogBox.body.innerHTML = `
+
+			<h2>Lets Get You Started</h2>
+
+			<a href="${demo_url}" target="_blank" class="view-demo">
+				<span class="figure"><i class="fas fa-external-link-alt"></i></span>
+				<span>View Demo</span>
+			</a>
+
+			<a class="initiate-walkthrough">
+				<span class="figure"><i class="fas fa-cogs"></i></span>
+				<span>Configure Manually</span>
+			</a>
+
+			<a class="skip">Skip &nbsp;<i class="fas fa-arrow-right"></i></a>
+		`;
+
+		this.dialogBox.body.querySelector('.initiate-walkthrough').on('click', () => this.dialogBox.hide);
+
+		if(window.loadWelcomeDialogBoxListener)
+			window.loadWelcomeDialogBoxListener(this);
+
+		this.dialogBox.body.querySelector('.skip').on('click', async () => {
+
+			this.dialogBox.hide();
+
+			const newUser = await Storage.get('newUser');
+
+			newUser.skipWelcomeDialogBox = true;
+
+			await Storage.set('newUser', newUser);
+		});
+
+		this.dialogBox.show();
 	}
 }
 
@@ -55,10 +131,8 @@ UserOnboard.stages.add(class AddConnection {
 
 	get container() {
 
-		if(this.containerElement) {
-
+		if(this.containerElement)
 			return this.containerElement;
-		}
 
 		const container = this.containerElement = document.createElement('div');
 		container.classList.add('stage');
@@ -100,17 +174,15 @@ UserOnboard.stages.add(class AddReport {
 
 	get container() {
 
-		if(this.containerElement) {
-
+		if(this.containerElement)
 			return this.containerElement;
-		}
 
 		const container = this.containerElement = document.createElement('div');
 		container.classList.add('stage');
 
 		container.innerHTML = `
 			<span class="order">2</span>
-			<span>Add Report</span>
+			<span>Create Report</span>
 		`;
 
 		container.on('click', () => {
@@ -145,17 +217,15 @@ UserOnboard.stages.add(class AddDashboard {
 
 	get container() {
 
-		if(this.containerElement) {
-
+		if(this.containerElement)
 			return this.containerElement;
-		}
 
 		const container = this.containerElement = document.createElement('div');
 		container.classList.add('stage');
 
 		container.innerHTML = `
 			<span class="order">3</span>
-			<span> Add Dashboard</span>
+			<span>Create Dashboard</span>
 		`;
 
 		container.on('click', () => {
@@ -190,17 +260,15 @@ UserOnboard.stages.add(class AddVisualization {
 
 	get container() {
 
-		if(this.containerElement) {
-
+		if(this.containerElement)
 			return this.containerElement;
-		}
 
 		const container = this.containerElement = document.createElement('div');
 		container.classList.add('stage');
 
 		container.innerHTML = `
 			<span class="order">4</span>
-			<span>Add Visualization</span>
+			<span>Create Visualization</span>
 		`;
 
 		container.on('click', () => {
@@ -232,7 +300,6 @@ UserOnboard.stages.add(class AddVisualization {
 			}
 		}
 	}
-
 });
 
 UserOnboard.setup();
